requirements.txt
selenium>=4.0.0
pdfplumber>=0.9.0
camelot-py>=0.11.0
xlwt>=1.3.0
requests>=2.28.0
